Move tuple implementation to separate module
