+++ numba/targets/base.py
@@ -18,7 +18,8 @@ from numba.targets.imputils import (user_function, user_generator,
                                     builtin_registry, impl_attribute,
                                     type_registry,
                                     impl_ret_borrowed)
-from . import arrayobj, builtins, iterators, rangeobj, optional, slicing
+from . import (
+    arrayobj, builtins, iterators, rangeobj, optional, slicing, tupleobj)
 from numba import datamodel
 
 try:
